+++ src/services/api.ts
@@ -2,10 +2,8 @@ import axios, {AxiosInstance, InternalAxiosRequestConfig, AxiosResponse, AxiosEr
 import {getToken} from './token.ts';
 import {StatusCodes} from 'http-status-codes';
 import {processErrorHandle} from './process-error-handle.ts';
-import { Offer } from '../types/Offer.ts';
-import { Cities } from '../const.ts';
 
-const BACKEND_URL = 'https://15.design.htmlacademy.pro/six-cities/offers';
+const BACKEND_URL = 'https://15.design.htmlacademy.pro/six-cities/';
 const REQUEST_TIMEOUT = 5000;
 
 type DetailMessageType = {
@@ -61,14 +59,3 @@ export const createAPI = (): AxiosInstance => {
   return api;
 };
 
-export const fetchOffersForCity = async (city: Cities): Promise<Offer[]> => {
-  const response = await fetch(`${BACKEND_URL}/api/offers?city=${city}`);
-
-  if (!response.ok) {
-    const errorText = await response.text();
-    console.error('Error fetching offers:', errorText);
-    throw new Error('Failed to fetch offers for the selected city');
-  }
-
-  return await response.json();
-};
